（再）家計簿一覧機能テスト（コントローラー）

+++ src/main/java/com/example/controller/kakeibo/KakeiboController.java
@@ -177,7 +177,7 @@ public class KakeiboController {
 	 * @return
 	 */
 	@PostMapping("/delete")
-	public String deleteKakeibo(Integer id) {
+	public String deleteKakeibo(long id) {
 		kakeiboService.deleteKakeibo(id);
 		
 		return "redirect:/kakeibo/list";
